Move agent.py to root directory
As requested by the user, this change moves the agent code from `src/human_learning_ME/agent.py` back to `agents.py` in the root directory to simplify the project structure.

The import statements in `simulations.py` have been updated to reflect this change. The `src` directory has been removed.
